Refactor data module into featureset, readers, and writers modules.
The main idea here was that I wanted to make it easier for us to write
utilities that manipulate feature sets.  I've updated skll_convert to
use the new Readers and Writers, but I should also be able to replace
join_megam and filter_megam with more generic join_feature_set and
filter_feature_set tools.

This is not entirely done yet, because I still need to refactor
the experiments module to use all of this stuff.