Replace public `SCORERS` with private `_SCORERS`.
Scikit-learn has modified their metrics interface to deprecate the
`SCORERS` dictionary and made it private instead so we need to start
using the private interface. This is a bit risky as the private
interface can change at any time but since we pin scikit-learn, this is
less of an issue.